Fix for new layout
Frankly, it just opens the old layout again, cause the new one doesn't seem to be controllable by script, everything is in iframes on a different domain, it seems.
For the same reason, you now have to manually click on the appropriate container, probably, to see the backup at the end.